fix bool handling + organize code
